Refactor `EchoMeasure` to be independent from CAMUS
- Moved outside of CAMUS package
- Do not use CAMUS labels by default
